Шаблонизируй это 2.0, дополненная и исправленная :)
