+++ src/model/films-model.js
@@ -1,7 +1,11 @@
 import {generateFilm} from '../mock/film.js';
 
 export default class FilmsModel {
-  films = Array.from({length: 5}, generateFilm);
+  #films = Array.from({length: 5}, generateFilm);
 
-  getFilms = () => this.films;
+  getFilms () {
+    return this.#films;
+  }
 }
+
+
