+++ src/view/popup-film-container-viev.js
@@ -1,0 +1,137 @@
+import {createElement} from '../render.js';
+import {humanizeDate} from '../utils.js';
+//import {generateFilm} from '../mock/film.js';
+
+const createPopupFilmContainerTemplate = (film) => {
+  //const film =  generateFilm();
+  const {
+    //id,
+    filmInfo:{
+      title,
+      alternativeTitle,
+      totalRating,
+      poster,
+      ageRating,
+      director,
+      writers,
+      actors,
+      release: {
+        date,
+        releaseCountry
+      },
+      runtime,
+      genre,
+      description
+    },
+    //userDetails:{
+    //watchlist,
+    //alreadyWatched,
+    //watchingDate,
+    //favoriten
+    //}
+  } = film;
+
+  const template = 'DD MMM YYYY';
+  const humanizedDate = date !== null
+    ? humanizeDate(date, template)
+    : '';
+
+  return ( `
+    <div class="film-details__top-container">
+      <div class="film-details__close">
+        <button class="film-details__close-btn" type="button">close</button>
+      </div>
+      <div class="film-details__info-wrap">
+        <div class="film-details__poster">
+          <img class="film-details__poster-img" src="./${poster}" alt="">
+            <p class="film-details__age">${ageRating}+</p>
+        </div>
+
+        <div class="film-details__info">
+          <div class="film-details__info-head">
+            <div class="film-details__title-wrap">
+              <h3 class="film-details__title">${title}</h3>
+              <p class="film-details__title-original">Original: ${alternativeTitle}</p>
+            </div>
+
+            <div class="film-details__rating">
+              <p class="film-details__total-rating">${totalRating}</p>
+            </div>
+          </div>
+
+          <table class="film-details__table">
+            <tr class="film-details__row">
+              <td class="film-details__term">Director</td>
+              <td class="film-details__cell">${director}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Writers</td>
+              <td class="film-details__cell">${writers}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Actors</td>
+              <td class="film-details__cell">${actors}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Release Date</td>
+              <td class="film-details__cell">${humanizedDate}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Runtime</td>
+              <td class="film-details__cell">${runtime}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Country</td>
+              <td class="film-details__cell">${releaseCountry}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Genres</td>
+              <td class="film-details__cell">
+                ${genre}
+              </td>
+            </tr>
+          </table>
+
+          <p class="film-details__film-description">
+            ${description}
+          </p>
+        </div>
+      </div>
+
+      <section class="film-details__controls">
+        <button type="button"
+                class="film-details__control-button film-details__control-button--active film-details__control-button--watchlist"
+                id="watchlist" name="watchlist">Add to watchlist
+        </button>
+        <button type="button" class="film-details__control-button film-details__control-button--watched" id="watched"
+                name="watched">Already watched
+        </button>
+        <button type="button" class="film-details__control-button film-details__control-button--favorite" id="favorite"
+                name="favorite">Add to favorites
+        </button>
+      </section>
+    </div>`
+  );
+};
+
+export default class PopupFilmsContainerView {
+  constructor(film) {
+    this.film = film;
+  }
+
+  getTemplate() {
+    return createPopupFilmContainerTemplate(this.film);
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
